+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added SSR and hydration support for initial `<ErrorBoundary>` children.
+- Added SSR and hydration support for initial `<Transition>` children and v-if chains.
 
 ## 1.0.20 - 2026-05-13